requirements.txt
Flask
requests
openai
python-dotenv
google-api-python-client